# internal/adapters/http/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for centralized API key management and service access control.",
        "title": "Pandora Core",
        "termsOfService": "http://example.com/terms/",
        "contact": {
            "name": "Pandora Core Support",
            "url": "http://example.com/support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/services": {
            "get": {
                "security": [
                    {
                        "OAuth2Password": []
                    }
                ],
                "description": "Fetches a list of all registered services",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Services"
                ],
                "summary": "Retrieves all services",
                "parameters": [
                    {
                        "enum": [
                            "active",
                            "deactivated",
                            "deprecated"
                        ],
                        "type": "string",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.ServiceResponse"
                            }
                        }
                    },
                    "default": {
                        "description": "Default error response for all failures",
                        "schema": {
                            "$ref": "#/definitions/errors.HTTPError"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "OAuth2Password": []
                    }
                ],
                "description": "Adds a new service to the system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Services"
                ],
                "summary": "Creates a new service",
                "parameters": [
                    {
                        "description": "Service creation data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ServiceCreate"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.ServiceResponse"
                        }
                    },
                    "default": {
                        "description": "Default error response for all failures",
                        "schema": {
                            "$ref": "#/definitions/errors.HTTPError"
                        }
                    }
                }
            }
        },
        "/api/v1/services/{id}": {
            "delete": {
                "security": [
                    {
                        "OAuth2Password": []
                    }
                ],
                "description": "Permanently removes a service by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Services"
                ],
                "summary": "Deletes a service",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Service ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "default": {
                        "description": "Default error response for all failures",
                        "schema": {
                            "$ref": "#/definitions/errors.HTTPError"
                        }
                    }
                }
            }
        },
        "/api/v1/services/{id}/status": {
            "patch": {
                "security": [
                    {
                        "OAuth2Password": []
                    }
                ],
                "description": "Changes the current status of a specific service by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Services"
                ],
                "summary": "Updates the status of a service",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Service ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New service status",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ServiceStatusUpdate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ServiceResponse"
                        }
                    },
                    "default": {
                        "description": "Default error response for all failures",
                        "schema": {
                            "$ref": "#/definitions/errors.HTTPError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "dto.ServiceCreate": {
            "type": "object",
            "required": [
                "name",
                "version"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "dto.ServiceResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "active",
                        "deactivated",
                        "deprecated"
                    ]
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "dto.ServiceStatusUpdate": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string",
                    "enum": [
                        "active",
                        "deactivated",
                        "deprecated"
                    ]
                }
            }
        },
...
        "errors.ErrorCode": {
            "type": "string",
            "enum": [
                "NOT_FOUND",
                "INTERNAL",
                "FORBIDDEN",
                "UNAUTHORIZED",
                "ALREADY_EXISTS",
                "VALIDATION_FAILED",
                "AGGREGATE_ERRORS"
            ],
            "x-enum-varnames": [
                "CodeNotFound",
                "CodeInternal",
                "CodeForbidden",
                "CodeUnauthorized",
                "CodeAlreadyExists",
                "CodeValidationFailed",
                "CodeAggregate"
            ]
        },
        "errors.HTTPError": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/errors.ErrorCode"
                },
                "entity": {
                    "type": "string"
                },
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/errors.HTTPError"
                    }
                },
                "identifiers": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "loc": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "OAuth2Password": {
            "type": "oauth2",
            "flow": "password",
            "tokenUrl": "/api/v1/auth/login"
        }
    },
...
}